🚚 move sync transaction out
