+++ packages/atom.io/realtime-server/src/hook-composition/sync-transaction.ts
@@ -1,0 +1,33 @@
+import * as AtomIO from "atom.io"
+import * as Internal from "atom.io/internal"
+
+import type { ServerConfig } from "."
+
+export function useSyncTransaction({
+	socket,
+	store = Internal.IMPLICIT.STORE,
+}: ServerConfig) {
+	return function syncTransaction<ƒ extends AtomIO.ƒn>(
+		tx: AtomIO.TransactionToken<ƒ>,
+	): () => void {
+		const fillTransactionRequest = (
+			update: AtomIO.TransactionUpdate<ƒ>,
+			transactionId: string,
+		) => {
+			const unsubscribe = Internal.subscribeToTransaction(
+				tx,
+				(update) => {
+					unsubscribe()
+					socket.emit(`tx:sync:${transactionId}`, update)
+				},
+				`sync:${transactionId}`,
+				store,
+			)
+			AtomIO.runTransaction<ƒ>(tx, store)(...update.params)
+		}
+
+		socket.on(`tx:${tx.key}`, fillTransactionRequest)
+
+		return () => socket.off(`tx:${tx.key}`, fillTransactionRequest)
+	}
+}
